requirements.txt
docx2pdf>=0.1.8
lxml==6.0.2
Pillow>=10.0.0
PyMuPDF>=1.24.0
python-docx==1.2.0
PyYAML==6.0.3
typing_extensions==4.15.0
